Adding support for Elegoo Centauri (#115)

+++ src/Overseer/Machines/RepRapFirmware/Models/ConnectResponse.cs
@@ -6,5 +6,5 @@ public class ConnectResponse
   public int Err { get; set; }
   public int SessionTimeout { get; set; }
   public string BoardType { get; set; }
-  public int SessionKey { get; set; }
+  public long SessionKey { get; set; }
 }
